+++ frontend/components/common/modal/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useModal } from './useModal';
