+++ frontend/components/filter/jobsFilter/hooks/index.ts
@@ -2,3 +2,4 @@ export { useSelectedCategory } from './useSelectedCategory';
 export { useCheckedDetails } from './useCheckedDetails';
 export { useSelectedJobs } from './useSelectedJobs';
 export { useAllJobs } from './useAllJobs';
+export { useJobs } from './useJobs';
