+++ frontend/components/filter/jobsFilter/hooks/useAllJobs.ts
@@ -1,14 +1,18 @@
+import { useSearchParams } from 'next/navigation';
+
 import { https } from '@/apis/fetch';
 
 import { ResponseJobs } from '@/types/api/jobs';
+import { isValidServiceType } from '@/types/guards/queryParams';
 
-import { JOBS_URL } from '@/constants/api';
+import { JOBS_URL, PARAMS } from '@/constants/api';
 
-import { useFetch } from '@/hooks/useFetch';
-import { useServiceType } from '@/hooks/useServiceType';
+import { useFetch } from '@/hooks';
 
 export const useAllJobs = () => {
-  const url = `${JOBS_URL}?serviceType=${useServiceType() ?? '전체'}`;
+  const searchParams = useSearchParams();
+  const serviceType = searchParams.get(PARAMS.SERVICE_TYPE);
+  const url = `${JOBS_URL}?serviceType=${isValidServiceType(serviceType) ? serviceType : '전체'}`;
   const { data, isLoading, error } = useFetch<string, ResponseJobs>({
     fetch: () => https.get<ResponseJobs>(url),
     key: url,
