+++ frontend/components/filter/locationFilter/components/index.ts
@@ -1,0 +1,6 @@
+export { ApplyLocationFilterButton } from './applyLocationFilterButton';
+export { LocationFilterModalHeader } from './locationFilterModalHeader';
+export { ResetLocationFilterButton } from './resetLocationFilterButton';
+export { LocationFilterModalOpenButton } from './locationFilterModalOpenButton';
+export { CitiesCheckbox } from './citiesCheckbox';
+export { DistrictSelect } from './districtSelect';
