+++ frontend/components/filter/locationFilter/components/resetLocationFilterButton.tsx
@@ -4,7 +4,7 @@ interface Props {
   clearFilterLocations: () => void;
 }
 
-export default function ResetButton({ clearFilterLocations }: Props) {
+export function ResetLocationFilterButton({ clearFilterLocations }: Props) {
   return (
     <Button
       className='py-1 px-4 rounded border border-default-color border-opacity-30 hover:bg-default-color hover:bg-opacity-10'
