+++ frontend/components/filter/locationFilter/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useFilterLocations } from './useFilterLocations';
+export { useLocations } from './useLocations';
