+++ frontend/components/recruitment/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useInitialRecruitment } from './useInitialRecruitment';
 export { useRecruitment } from './useRecruitment';
+export { usePage } from './usePage';
