--- frontend/hooks/useEducationLevel.ts
@@ -1,12 +1,0 @@
-import { useSearchParams } from 'next/navigation';
-
-import { isValidEducationLevel } from '@/types/guards/queryParams';
-
-import { PARAMS } from '@/constants/api';
-
-export const useEducationLevel = () => {
-  const searchParams = useSearchParams();
-  const educationLevel = searchParams.get(PARAMS.EDUCATION_LEVEL);
-
-  return isValidEducationLevel(educationLevel) ? educationLevel : null;
-};
