--- frontend/hooks/useServiceStatus.ts
@@ -1,12 +1,0 @@
-import { useSearchParams } from 'next/navigation';
-
-import { isValidServiceStatus } from '@/types/guards/queryParams';
-
-import { PARAMS } from '@/constants/api';
-
-export const useServiceStatus = () => {
-  const searchParams = useSearchParams();
-  const serviceStatus = searchParams.get(PARAMS.SERVICE_STATUS);
-
-  return isValidServiceStatus(serviceStatus) ? serviceStatus : null;
-};
