--- frontend/hooks/useServiceType.ts
@@ -1,12 +1,0 @@
-import { useSearchParams } from 'next/navigation';
-
-import { isValidServiceType } from '@/types/guards/queryParams';
-
-import { PARAMS } from '@/constants/api';
-
-export const useServiceType = () => {
-  const searchParams = useSearchParams();
-  const serviceType = searchParams.get(PARAMS.SERVICE_TYPE);
-
-  return isValidServiceType(serviceType) ? serviceType : null;
-};
