--- frontend/hooks/useSort.ts
@@ -1,12 +1,0 @@
-import { useSearchParams } from 'next/navigation';
-
-import { isValidSort } from '@/types/guards/queryParams';
-
-import { PARAMS } from '@/constants/api';
-
-export const useSort = () => {
-  const searchParams = useSearchParams();
-  const sort = searchParams.get(PARAMS.SORT);
-
-  return isValidSort(sort) ? sort : null;
-};
